shop/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
psycopg2-binary
alembic
pytest
httpx
passlib[bcrypt]
PyJWT
pydantic-settings
